requirements.txt
# django
Django==3.1.3
pytz==2020.4
sqlparse==0.4.1

# djangorestframework
djangorestframework==3.12.2

# Pillow
Pillow==8.0.1

# requests-oauthlib
certifi==2020.11.8
chardet==3.0.4
idna==2.10
oauthlib==3.1.0
requests==2.25.0
requests-oauthlib==1.3.0
urllib3==1.26.2

# python-dotenv
python-dotenv==0.15.0
